+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 Versions follow [Semantic Versioning](https://www.semver.org)
 
+## [1.0.0] 2018-12-04
+
+[#8](https://github.com/T4rk1n/pytest-dash/pull/8)
+
+### Added
+
+- Added `dash_subprocess` fixture, runs a dash app in a subprocess waitress-serve command.
+- `utils.wait_for_text_to_equal`
+- `utils.wait_for_element_by_css_selector`
+
+### Removed
+- `dash_app` fixture.
+
+### Renamed
+- `start_dash` fixture to `dash_threaded`
+
 ## [0.1.3] 2018-10-04
 ### Fixed